+++ e2e/commands/commit.e2e.js
@@ -535,7 +535,9 @@ describe('bit tag command', function () {
     });
     it('should not tag and throw an error regarding the relative syntax', () => {
       expect(output).to.have.string('error: issues found with the following component dependencies');
-      expect(output).to.have.string('relative components (should be absolute):');
+      expect(output).to.have.string(
+        'components with relative import statements (please use absolute paths for imported components)'
+      );
       expect(output).to.have.string(`${helper.remoteScope}/utils/is-type@0.0.1`);
     });
   });