+++ specs/jsdoc/fixtures/endWithTwoStars.js
@@ -1,0 +1,9 @@
+/**
+ * Basic accordion component
+ * @render react
+ * @example
+ * <div>
+ *  <StoreList data={[{summary:"Item", details:"Convallis qui non phasellus nullam congue et pellentesque eros"}]} />
+ * </div>
+**/
+export default class Accordion extends React.Component { }