+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fix catastrophic backtracking when using Regex to find JSDoc
 - enable running build/test from internal directory of a scope
 - fix exit status when ci-update fails